job/pyproject.toml
[tool.poetry]
name = "TokaiOnAir-Video-Search-JOB"
version = "0.1.0"
description = ""
authors = ["nomnomnonono <nozomu.koujiguchi1204@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0.0"
google-api-python-client = "^2.101.0"
pandas = "^2.1.1"
google-cloud-storage = "^2.11.0"
cloud-sql-python-connector = "^1.4.1"
SQLAlchemy = "^2.0.21"
pymysql = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
